refactor(design-system): implement tailwind config package
