+++ packages/web/design-system/src/index.css
@@ -1,0 +1,1 @@
+@import '@wisemen/vue-core-tailwind-config';
