+++ packages/web/design-system/src/index.ts
@@ -1,3 +1,6 @@
+import './source.css'
+import './utilities.css'
+
 export * from './locales'
 export type { Register } from './register'
 export { type HourCycle } from './types/hourCycle.type'
