+++ packages/web/design-system/src/source.css
@@ -1,0 +1,1 @@
+@source "./";
