--- packages/web/design-system/src/styles/index.css
@@ -1,5 +1,0 @@
-@import './themes/default.css';
-@import './core/colors.css';
-@import './core/tokens.css';
-@import 'tailwindcss';
-@import './utilities.css';
