--- packages/web/design-system/src/styles/source.css
@@ -1,1 +1,0 @@
-@source "./";
