+++ packages/web/design-system/src/ui/dialog/dialog.style.ts
@@ -1,6 +1,6 @@
 import type { VariantProps } from 'tailwind-variants'
 
-import { tv } from '@/styles/tailwindVariants.lib'
+import { tv } from '@/libs/tailwindVariants.lib'
 
 export const createDialogStyle = tv({
   slots: {
@@ -9,8 +9,8 @@ export const createDialogStyle = tv({
       group-has-data-dialog-footer/dialog:pb-xs
     `,
     chin: `
-      absolute bottom-0 -z-10 w-full rounded-t-[calc(1rem+5px)] rounded-b-none
-      shadow-xl transition-transform duration-200
+      shadow-xl absolute bottom-0 -z-10 w-full rounded-t-[calc(1rem+5px)]
+      rounded-b-none transition-transform duration-200
       sm:top-0 sm:rounded-[calc(1rem+5px)]
     `,
     content: `
@@ -20,7 +20,7 @@ export const createDialogStyle = tv({
       sm:max-h-[85vh] sm:rounded-[calc(1rem+5px)] sm:border-[5px]
     `,
     contentWrapper: `
-      fixed inset-x-0 bottom-0 z-modal flex w-full flex-col
+      fixed inset-x-0 bottom-0 z-40 flex w-full flex-col
       will-change-[transform,opacity] outline-none
       sm:inset-x-auto sm:top-1/2 sm:bottom-auto sm:left-1/2 sm:-translate-1/2
     `,
@@ -28,7 +28,7 @@ export const createDialogStyle = tv({
     header: 'sticky top-0 z-10 bg-primary',
     innerContent: '',
     overlay: `
-      fixed inset-0 z-modal bg-linear-to-t from-black/50 to-black/25
+      fixed inset-0 z-40 bg-linear-to-t from-black/50 to-black/25
       will-change-[opacity]
       dark:from-black/80 dark:to-black/50
     `,
