+++ packages/web/design-system/src/ui/dot/dot.style.ts
@@ -1,4 +1,4 @@
-import { tv } from '@/styles/tailwindVariants.lib'
+import { tv } from '@/libs/tailwindVariants.lib'
 
 export const dotVariants = tv({
   base: 'shrink-0 rounded-full',
