+++ packages/web/design-system/src/ui/separator/separator.style.ts
@@ -1,4 +1,4 @@
-import { tv } from '@/styles/tailwindVariants.lib'
+import { tv } from '@/libs/tailwindVariants.lib'
 
 export const createSeparatorStyle = tv({
   slots: {
