+++ packages/web/design-system/src/utilities.css
@@ -1,45 +1,7 @@
-@variant dark (&:where(.dark, .dark *));
-
-@utility z-sticky {
-  z-index: 10;
-}
-
-@utility z-overlay {
-  z-index: 30;
-}
-
-@utility z-modal {
-  z-index: 40;
-}
-
-@utility z-tooltip {
-  z-index: 50;
-}
-
-@utility no-scrollbar {
-  &::-webkit-scrollbar {
-    display: none;
-  }
-  -ms-overflow-style: none;
-  scrollbar-width: none;
-}
-
-@utility glassy {
-  background:
-    linear-gradient(325deg, transparent 0%, transparent 80%, rgba(255, 255, 255, 0.4) 100%),
-    linear-gradient(45deg, transparent 0%, transparent 85%, rgba(0, 0, 0, 0.8) 100%),
-    linear-gradient(150deg, transparent 0%, transparent 80%, rgba(255, 255, 255, 0.4) 100%),
-    linear-gradient(225deg, transparent 0%, transparent 85%, rgba(0, 0, 0, 0.8) 100%), rgba(255, 255, 255, 0.4);
-}
-
 @utility shadow-floating {
   box-shadow:
     0px 3px 3px -1.5px #3641520a,
     0px 8px 8px -4px #36415208;
-}
-
-@utility glassy-inner-content {
-  @apply bg-linear-to-br from-secondary to-quaternary group-hover:from-primary;
 }
 
 @keyframes popoverFadeInFromTop {
