+++ packages/web/filters/src/index.css
@@ -1,2 +1,1 @@
 @import '@wisemen/vue-core-tailwind-config';
-@source './';
